EUREKA!
So, my algorithm for finding edges allows turning back, in addition to
proceeding straight (if we haven't gone straight too far) or turning
left or right.

That's not supposed to be allowed, but, if it is allowed, I successfully
found a path with a cost of 101 instead of 102 by using a "double-back"
(which at first blush would seem could only be counter-productive) to
break-up an otherwise-disallowed sequence of 5 consecutive "south"
moves!